Rename group name to com.teletronics.notesapi
