--- src/main/java/com/tele/notesapi/NotesApiApplication.java
@@ -1,13 +1,0 @@
-package com.tele.notesapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotesApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NotesApiApplication.class, args);
-	}
-
-}
