--- src/main/java/com/tele/notesapi/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.tele.notesapi.exception;
-
-public class ResourceNotFoundException extends RuntimeException{
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
