--- src/main/java/com/tele/notesapi/exception/handler/GlobalExceptionHandler.java
@@ -1,54 +1,0 @@
-package com.tele.notesapi.exception.handler;
-
-import com.tele.notesapi.exception.ResourceNotFoundException;
-import com.tele.notesapi.model.ErrorResponse;
-import jakarta.validation.ConstraintViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.method.annotation.HandlerMethodValidationException;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-
-import java.time.LocalDateTime;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler({
-            ConstraintViolationException.class,
-            MethodArgumentNotValidException.class,
-            HandlerMethodValidationException.class,
-            MethodArgumentTypeMismatchException.class
-    })
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ResponseEntity<ErrorResponse> handleValidationExceptions(Exception ex) {
-        ErrorResponse errorResponse = createErrorResponse(HttpStatus.BAD_REQUEST, "Validation error", ex.getMessage());
-        return ResponseEntity.badRequest().body(errorResponse);
-    }
-
-    @ExceptionHandler(ResourceNotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ResponseEntity<ErrorResponse> handleResourceNotFoundException(ResourceNotFoundException ex) {
-        ErrorResponse errorResponse = createErrorResponse(HttpStatus.NOT_FOUND, "Resource not found", ex.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
-    }
-
-    @ExceptionHandler(Exception.class)
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public ResponseEntity<ErrorResponse> handleException(Exception ex) {
-        ErrorResponse errorResponse = createErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, "Internal server error", ex.getMessage());
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-    }
-
-    private ErrorResponse createErrorResponse(HttpStatus status, String error, String message) {
-        ErrorResponse errorResponse = new ErrorResponse();
-        errorResponse.setTimestamp(LocalDateTime.now());
-        errorResponse.setStatus(status.value());
-        errorResponse.setError(error);
-        errorResponse.setMessage(message);
-        return errorResponse;
-    }
-}
