--- src/main/java/com/tele/notesapi/model/Constant.java
@@ -1,10 +1,0 @@
-package com.tele.notesapi.model;
-
-public class Constant {
-
-    public enum Tag {
-        BUSINESS,
-        PERSONAL,
-        IMPORTANT
-    }
-}
