--- src/main/java/com/tele/notesapi/model/ErrorResponse.java
@@ -1,15 +1,0 @@
-package com.tele.notesapi.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-public class ErrorResponse {
-    private LocalDateTime timestamp;
-    private int status;
-    private String error;
-    private String message;
-}
