--- src/main/java/com/tele/notesapi/model/NoteResponse.java
@@ -1,24 +1,0 @@
-package com.tele.notesapi.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-import java.util.Set;
-import java.util.UUID;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@RequiredArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class NoteResponse {
-    private UUID id;
-    private String title;
-    private String text;
-    private Set<Constant.Tag> tags;
-    private LocalDateTime createdDate;
-}
