--- src/main/java/com/tele/notesapi/model/NoteSummary.java
@@ -1,17 +1,0 @@
-package com.tele.notesapi.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class NoteSummary {
-    private String title;
-    private LocalDateTime createdDate;
-}
