+++ src/main/java/com/teletronics/notesapi/NotesApiApplication.java
@@ -1,0 +1,13 @@
+package com.teletronics.notesapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NotesApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NotesApiApplication.class, args);
+	}
+
+}
