+++ src/main/java/com/teletronics/notesapi/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.teletronics.notesapi.exception;
+
+public class ResourceNotFoundException extends RuntimeException{
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
