+++ src/main/java/com/teletronics/notesapi/model/Constant.java
@@ -1,0 +1,10 @@
+package com.teletronics.notesapi.model;
+
+public class Constant {
+
+    public enum Tag {
+        BUSINESS,
+        PERSONAL,
+        IMPORTANT
+    }
+}
