+++ src/main/java/com/teletronics/notesapi/model/ErrorResponse.java
@@ -1,0 +1,15 @@
+package com.teletronics.notesapi.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+public class ErrorResponse {
+    private LocalDateTime timestamp;
+    private int status;
+    private String error;
+    private String message;
+}
