+++ src/main/java/com/teletronics/notesapi/model/NoteSummary.java
@@ -1,0 +1,17 @@
+package com.teletronics.notesapi.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@RequiredArgsConstructor
+public class NoteSummary {
+    private String title;
+    private LocalDateTime createdDate;
+}
