--- src/test/java/com/tele/notesapi/NotesApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tele.notesapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NotesApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
