+++ src/test/java/com/teletronics/notesapi/NotesApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.teletronics.notesapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotesApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
